#34 +Result{InitialVulnerability, ContextualVulnerability, RefinedVulnerability}
In the exported RESULT, the Vulnerability is calculated for each ThreatAgent & for each phase.
This way it is possible to extract from it all the remaining values.
E.G.
InitialLikelihood = MAX(InitialVulnerability)
ContextualLikelihood = MAX(ContextualVulnerability)
RefinedLikelihood = MAX(RefinedVulnerability)

TODO: Fix the calculation of the likelihood.
At the moment, for the denominator of the MainFormula, only the weight of the phases where there is a feasible AttackStrategy by the current ThreatAgent, are used in the sum, but indeed the denominator should be the sum of the weights of all phases, regardless of the Attacks feasible by the ThreatAgent