-Fix #47
The initial approach was to show to the user the AttackMap updated in real time while answering the questionnaire, then replace it with the result of the questionnaire at the end, but as better explained in issue #45, it requires additional work. So, we simply show only the final result of the AM matrix.