+Fix #196: @ControllerAdvice + @Priority
Having two @ControllerAdvice what happened was that (at least in production, but not in dev profile) the ZalandoExceptionTranslator was called first and not having an handler for our custom exceptions it was a 500 internal server error, instead of 404.
With @Priority the GlobalExceptionHandler gets called first and can handle the exception response with 404.

- https://books.google.it/books?id=Rr1OwYRj8XMC&pg=PA53&lpg=PA53&dq=javax.annotation.Priority+which+value+comes+first&source=bl&ots=HGN3V-KSWv&sig=cTF-g0zn7ojXbIh6lhRnjzoNN5s&hl=en&sa=X&ved=2ahUKEwid2bezpa_fAhWP66QKHf-3APwQ6AEwBnoECAAQAQ#v=onepage&q=javax.annotation.Priority%20which%20value%20comes%20first&f=false

- https://stackoverflow.com/questions/19498378/setting-precedence-of-multiple-controlleradvice-exceptionhandlers

- https://www.baeldung.com/global-error-handler-in-a-spring-rest-api